refactor: restructure repo to reflect monorepo with two services
- Reorganize into three top-level folders:
  - back-end/: Go service with its own module
  - front-end/: React frontend (renamed from web/)
  - acceptance/: BDD tests with own module importing from back-end/pkg/
- Move domain entities to back-end/pkg/ for external access
- Create public wrappers for application and HTTP services
- Update all import paths to reflect new structure
- Add node_modules to .gitignore and remove from repo
- Update documentation to reflect new monorepo structure

Closes #8
