+++ README.md
@@ -4,11 +4,17 @@
 
 Demonstration of the Screenplay Pattern using BDD with Cucumber/Gherkin in Go.
 
-Also demonstrates how the same BDD scenarios can be run against different deployment models:
+This monorepo contains three main components:
+- **back-end**: Go service with domain logic and HTTP API
+- **front-end**: React frontend (formerly web/)
+- **acceptance**: BDD acceptance tests using Cucumber/Gherkin
+
+The same BDD scenarios can be run against different deployment models:
 - Direct domain access
 - HTTP API with in-process server
 - HTTP API with separate server executable
 - HTTP API with Docker container
+- Full UI testing with frontend and API in containers
 
 Based on the official [Cucumber Screenplay Example](https://github.com/cucumber-school/screenplay-example/tree/code) to `go`, using the official [godog](https://github.com/cucumber/godog/) library.
 
@@ -48,14 +54,15 @@ make test-all                 # Full test suite including Docker
 ### Direct Go Commands
 
 ```sh
-# Run all tests
-go test -v ./acceptance
+# Run all tests (from acceptance directory)
+cd acceptance && go test -v .
 
 # Run specific test types
-go test -v -run TestApplication ./acceptance
-go test -v -run TestHTTPInProcess ./acceptance
-go test -v -run TestHttpExecutable ./acceptance
-go test -v -run TestHttpDocker ./acceptance
+cd acceptance && go test -v -run TestApplication .
+cd acceptance && go test -v -run TestHTTPInProcess .
+cd acceptance && go test -v -run TestHttpExecutable .
+cd acceptance && go test -v -run TestHttpDocker .
+cd acceptance && go test -v -run TestUI .
 ```
 
 
@@ -73,26 +80,39 @@ There are some differences in structure:
 - `go` does not support arrow functions so the implementation of actions, tasks etc uses standard functions
 
 - to promote separation of concerns:
-   - the domain implementation code is placed in the `internal/domain` package following Go conventions
-   - the HTTP server implementation is in the `internal/http` package
-   - test drivers in `acceptance/driver` provide different ways to access the domain (direct, HTTP client, etc.)
+   - the domain implementation code is placed in the `back-end/internal/domain` package following Go conventions
+   - public domain interfaces are exposed via `back-end/pkg/domain/` for use by acceptance tests
+   - the HTTP server implementation is in the `back-end/internal/http` package
+   - test drivers in `acceptance/driver` provide different ways to access the domain (direct, HTTP client, UI automation)
    - We inject the application into the test suite via the go test functions so we no longer have an exported InitializeScenarios function. This means the tests can no longer be run from `godog run` but instead should be run from `go test`
-   - acceptance test code has been placed in the `acceptance` folder and split into several files
+   - acceptance test code has been placed in the `acceptance` folder with its own Go module and split into several files
 
 ## Architecture
 
-The project follows clean architecture principles:
+The project follows clean architecture principles with a monorepo structure:
 
 ```
-acceptance/         # BDD tests and test drivers
-├── driver/
+back-end/           # Go service (independent module)
+├── cmd/server/     # Runnable HTTP server
+├── internal/       # Internal implementation packages
+│   ├── domain/     # Core business logic
+│   └── http/       # HTTP server implementation
+└── pkg/            # Public packages for external use
+    ├── domain/     # Domain entities and services
+    └── http/       # HTTP server interface
+
+front-end/          # React frontend (formerly web/)
+├── src/            # React source code
+├── public/         # Static assets
+└── Dockerfile      # Frontend container
+
+acceptance/         # BDD tests (independent module)
+├── driver/         # Test drivers for different deployment modes
 │   ├── application/# Direct domain access driver
-│   └── http/       # HTTP client driver
+│   ├── http/       # HTTP client driver
+│   └── ui/         # UI automation driver
 ├── screenplay/     # Screenplay pattern implementation
-internal/           # Internal implementation packages
-├── domain/         # Core business logic
-└── http/           # HTTP server implementation
-cmd/server/         # Runnable HTTP server
+└── features/       # Gherkin feature files
 ```
 
 ## Test Levels
@@ -101,6 +121,7 @@ cmd/server/         # Runnable HTTP server
 - **HTTP In-Process** (`make test-http-inprocess`): HTTP API testing with in-process server (~4-5ms)
 - **Server Executable** (`make test-http-executable`): Full integration with separate server process (~1-2s)
 - **Docker Container** (`make test-http-docker`): Production-like containerized testing (~30-60s)
+- **UI Tests** (`make test-ui`): Full stack testing with frontend and API containers using browser automation (~60-120s)
 
 All tests run identical BDD scenarios ensuring contract compliance across all deployment models.
 
@@ -126,6 +147,19 @@ make build
 # Build and run server
 make server
 
-# Or run directly
-go run ./cmd/server
+# Or run directly from back-end directory
+cd back-end && go run ./cmd/server
+
+# Run frontend development server
+cd front-end && npm start
 ```
+
+## Module Structure
+
+Each component is now its own Go module for better dependency management:
+
+- `back-end/go.mod` - Backend service module
+- `acceptance/go.mod` - Acceptance tests module that imports backend as dependency
+- `front-end/package.json` - Frontend dependencies
+
+The acceptance tests use the backend's public API via `back-end/pkg/` packages.
