+++ acceptance/driver/http/http.go
@@ -8,7 +8,7 @@ import (
 	"net/http"
 
 	"github.com/sirockin/cucumber-screenplay-go/acceptance/driver"
-	"github.com/sirockin/cucumber-screenplay-go/internal/domain/entities"
+	"github.com/sirockin/cucumber-screenplay-go/back-end/pkg/domain/entities"
 )
 
 type AcceptanceTestDriver struct {
