+++ back-end/go.mod
@@ -1,0 +1,5 @@
+module github.com/sirockin/cucumber-screenplay-go/back-end
+
+go 1.24.0
+
+exclude google.golang.org/genproto v0.0.0-20191108220845-16a3f7862a1a
