+++ front-end/src/components/SignUp.js
@@ -1,0 +1,59 @@
+import React, { useState } from 'react';
+import { useNavigate } from 'react-router-dom';
+
+function SignUp() {
+  const [name, setName] = useState('');
+  const [message, setMessage] = useState('');
+  const [error, setError] = useState('');
+  const navigate = useNavigate();
+
+  const handleSubmit = async (e) => {
+    e.preventDefault();
+    setMessage('');
+    setError('');
+
+    try {
+      const response = await fetch('/accounts', {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify({ name }),
+      });
+
+      if (response.ok) {
+        setMessage(`Account created successfully for ${name}!`);
+        setTimeout(() => {
+          navigate(`/account/${name}`);
+        }, 1500);
+      } else {
+        const errorData = await response.text();
+        setError(`Failed to create account: ${errorData}`);
+      }
+    } catch (err) {
+      setError(`Network error: ${err.message}`);
+    }
+  };
+
+  return (
+    <div>
+      <h2>Sign Up</h2>
+      {message && <div className="success message">{message}</div>}
+      {error && <div className="error">{error}</div>}
+
+      <form onSubmit={handleSubmit} className="form">
+        <input
+          type="text"
+          name="name"
+          placeholder="Enter your name"
+          value={name}
+          onChange={(e) => setName(e.target.value)}
+          required
+        />
+        <button type="submit">Create Account</button>
+      </form>
+    </div>
+  );
+}
+
+export default SignUp;
