--- web/src/components/Login.js
@@ -1,55 +1,0 @@
-import React, { useState } from 'react';
-import { useNavigate } from 'react-router-dom';
-
-function Login() {
-  const [name, setName] = useState('');
-  const [message, setMessage] = useState('');
-  const [error, setError] = useState('');
-  const navigate = useNavigate();
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-    setMessage('');
-    setError('');
-
-    try {
-      const response = await fetch(`/accounts/${name}/authenticate`, {
-        method: 'POST',
-      });
-
-      if (response.ok) {
-        setMessage(`Successfully authenticated ${name}!`);
-        setTimeout(() => {
-          navigate(`/account/${name}`);
-        }, 1500);
-      } else {
-        const errorData = await response.json();
-        setError(errorData.error || 'Authentication failed');
-      }
-    } catch (err) {
-      setError(`Network error: ${err.message}`);
-    }
-  };
-
-  return (
-    <div>
-      <h2>Login</h2>
-      {message && <div className="success">{message}</div>}
-      {error && <div className="error">{error}</div>}
-
-      <form onSubmit={handleSubmit} className="form">
-        <input
-          type="text"
-          name="name"
-          placeholder="Enter your name"
-          value={name}
-          onChange={(e) => setName(e.target.value)}
-          required
-        />
-        <button type="submit">Login</button>
-      </form>
-    </div>
-  );
-}
-
-export default Login;
