--- web/src/index.css
@@ -1,135 +1,0 @@
-body {
-  margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-    sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-.container {
-  max-width: 800px;
-  margin: 0 auto;
-  padding: 20px;
-}
-
-.nav {
-  background-color: #f5f5f5;
-  padding: 10px;
-  margin-bottom: 20px;
-}
-
-.nav a {
-  margin-right: 15px;
-  text-decoration: none;
-  color: #333;
-}
-
-.nav a:hover {
-  text-decoration: underline;
-}
-
-.form {
-  background-color: #f9f9f9;
-  padding: 20px;
-  border-radius: 5px;
-  margin-bottom: 20px;
-}
-
-.form input {
-  width: 100%;
-  padding: 10px;
-  margin-bottom: 10px;
-  border: 1px solid #ddd;
-  border-radius: 3px;
-}
-
-.form button {
-  background-color: #007bff;
-  color: white;
-  padding: 10px 20px;
-  border: none;
-  border-radius: 3px;
-  cursor: pointer;
-}
-
-.form button:hover {
-  background-color: #0056b3;
-}
-
-.success {
-  background-color: #d4edda;
-  color: #155724;
-  padding: 10px;
-  border-radius: 3px;
-  margin-bottom: 10px;
-}
-
-.error {
-  background-color: #f8d7da;
-  color: #721c24;
-  padding: 10px;
-  border-radius: 3px;
-  margin-bottom: 10px;
-}
-
-.account-info {
-  background-color: #e9ecef;
-  padding: 15px;
-  border-radius: 5px;
-  margin-bottom: 20px;
-}
-
-.status-activated,
-.status-authenticated {
-  color: #28a745;
-  font-weight: bold;
-}
-
-.projects-list {
-  list-style: none;
-  padding: 0;
-}
-
-.project-item {
-  background-color: #f8f9fa;
-  padding: 10px;
-  margin-bottom: 5px;
-  border-radius: 3px;
-  border-left: 4px solid #007bff;
-}
-
-.activate {
-  background-color: #28a745;
-  color: white;
-  padding: 10px 20px;
-  border: none;
-  border-radius: 3px;
-  cursor: pointer;
-}
-
-.activate:hover {
-  background-color: #218838;
-}
-
-.create-project {
-  background-color: #17a2b8;
-  color: white;
-  padding: 10px 20px;
-  border: none;
-  border-radius: 3px;
-  cursor: pointer;
-  margin-bottom: 20px;
-}
-
-.create-project:hover {
-  background-color: #138496;
-}
-
-.project-created {
-  background-color: #d1ecf1;
-  color: #0c5460;
-  padding: 10px;
-  border-radius: 3px;
-  margin-bottom: 10px;
-}
